channelMask match number of channels
Fixed the bug with th extensible format channelMask validation.
The bug was that when creating BW64 files with the extensible format, the channelMask in the ExtraData was not being validated to ensure it has exactly the same number of bits set as the number of channels. This violates the WAV extensible format specification.